Pipeline for analysing gene-sets
The purpose of these scripts is to help better understand partitioning
by experimenting with and comparing various ways of partitioning gene
datasets. In particular, they help compare the speed and accuracy of
three tree-generation algorithms – RaxML, PhyloBayes and ‘Tree
Collection’, the last one having been created by Jeremy Levy. The first
two use probability models to guess the phylogenies most representative
of the observed data, the last uses distance to class genes into
phylogenies. To compare the accuracy of the algorithms, the results
from each will be subjected to a monophyly test. A monophyly test
assesses the extent to which trees correctly categorise the species
involved, e.g. making sure flies are with flies, rodents with rodents
etc. By using these tests of speed and accuracy, not just will we have
an idea of the best algorithm, but also of the way in which genes
should be partitioned - according to distances, or likelihood of
matching certain models? These models may simply be based on branch
length or the types of evolutionary change over time.